+++ litho-picasso/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.github.charbgr.litho.picasso">
+    package="com.github.charbgr.litho.picassox">
 </manifest>
